[#1742] Move StaticMapping's config schema to separate file
